+++ server/server.py
@@ -1,11 +1,10 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-import httpx
 import asyncio
 import requests
 import aiohttp
 from aiohttp_client_cache import CachedSession, SQLiteBackend
-
+from routes import recipe_router
 from dotenv import load_dotenv
 import os
 
@@ -54,21 +53,7 @@ app.add_middleware(
 
 # use beaker for caching
 
-@app.get('/')
-# this whole thing returns a coroutine. how do I have it actually run it..?
-async def basic_get():
-    # make api call
-    async with httpx.AsyncClient() as client:
-        r = await client.get(
-            'https://api.spoonacular.com/recipes/complexSearch', 
-            params={
-                "apiKey": SPOONACULAR_API_KEY,
-                "query": "tacos"
-            }
-        )
-    # data = asyncio.run(r)
-    # return info
-    return {'data': r.json()} # json to serialize it upon return
+
 
 
 async def get_api(session, url, params):
